app/requirements.txt
fastapi 
uvicorn
psycopg2-binary
python-jose[cryptography]
requests